Update finaid page with information for EP 2023
As finaid has two rounds of selection this year, two SVG-icon files have
been added to provide a visual accent for the timeline rows belonging to
either round 1 or round 2.

Note that the footnote for the deadline of round 2 does not actually
work properly at this moment. This is due to a missing anchor or id in
the rendering of the actual footnote at the bottom of the page. This
problem is probably related to the lack of anchor ids on header tags, as
described in #237.