Make round image legible on small displays
I've made some minor CSS-tweaks to make the image for the selection
round in the timeline legible on small screens.

In addition, I've removed the top/bottom padding on the images to
prevent inconsistent vertical spacing; not all lines in the timeline
have an image. Instead, I've added a bottom-padding to all table cells
to create space between table rows.