Add content pages to tailwind configuration
Tailwind automatically detects the tailwind classes used in files found
using the glob patterns in its "content" configuration. If a classname
is not used in a file covered by these globs, it won't be included in
the final build CSS-file.

As the content pages (.md) were not scanned by tailwind, if classname
was only used in such a file, it would not be included in the final
built CSS-file. This meant that only classes that were already in use
could be used in these content pages.

I've added a glob that ensures that tailwind scans the content files.