Tickets page second version  (#1045)
I created a new branch from the initial one because I wanted to avoid
removing things there.

Changes:  
* Rewrote it in Markdown for better readability  
* Removed a note about remote tickets  
* Added more features to the list  
* Rephrased some text  
* Used the same style as the rest of the page  

We agreed that it would be nice to have bigger headers and a wider page,
just like in the first MR. However, since that requires additional CSS
tweaks, I left it as a nice-to-have for later.

---------

Co-authored-by: Artur Czepiel <czepiel.artur@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Theofanis Petkos <thepetk@gmail.com>
Co-authored-by: Cristián Maureira-Fredes <cmaureir@users.noreply.github.com>