Standardise summit pages (#1054)
The three summit pages are similar, but differently formatted:

* https://ep2025.europython.eu/programme/c-api-summit/
* https://ep2025.europython.eu/programme/wasm-summit/
* https://ep2025.europython.eu/programme/rust-summit/

First, make sure all have standard headers (h1 for first-level, h2 for
second-level, and so on; not h2 for first-level and h4 for second-level)
- this is partly an accessibility thing.

Then give all the same type of when/where/who sections at the top.

And move the intro summit before that.

---------

Co-authored-by: Mia Bajić <38294198+clytaemnestra@users.noreply.github.com>